distribute small wrappers in main and split the remainder
